requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.2.9
django-cleanup==4.0.0
django-crispy-forms==1.8.1
django-extensions==2.2.9
graphviz==0.14
idna==2.9
joblib==0.15.1
numpy==1.18.4
pandas==1.0.3
Pillow==7.0.0
pprint==0.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.14.0
spotipy==2.12.0
sqlparse==0.3.0
threadpoolctl==2.0.0
urllib3==1.25.8
